+++ libs/components/src/menu/index.ts
@@ -1,5 +1,5 @@
 export * from "./menu.module";
 export * from "./menu.component";
 export * from "./menu-trigger-for.directive";
-export * from "./menu-item.directive";
+export * from "./menu-item.component";
 export * from "./menu-divider.component";
