+++ libs/components/src/menu/menu.component.ts
@@ -10,7 +10,7 @@ import {
   contentChildren,
 } from "@angular/core";
 
-import { MenuItemComponent } from "./menu-item.directive";
+import { MenuItemComponent } from "./menu-item.component";
 
 // FIXME(https://bitwarden.atlassian.net/browse/CL-764): Migrate to OnPush
 // eslint-disable-next-line @angular-eslint/prefer-on-push-component-change-detection
